% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ajdbmulti.R
\name{mtrmulti}
\alias{mtrmulti}
\title{Adjacency Matrix and Boolean Matrix Multiplication}
\usage{
mtrmulti(mtr, mtr2 = mtr, n = 2, type = 1)
}
\arguments{
\item{mtr}{squared matrix or proper matrix}

\item{mtr2}{proper matrix where dim(mtr)\link{2} == dim(mtr2)\link{1}}

\item{n}{power}

\item{type}{calculate adjacency matrix if 1, and boolean matrix if 2}
}
\value{
returns multiplied (or powered) adjacency matrix if type chosen was 1,
and boolean matrix if type chosen was 2
}
\description{
This function calculates adjacency matrix or boolean matrix
with desired power, "n". This function can have 2 matrices
to multiply rather than 1 matrix.
}
\note{
although the speed of the algorithm of adjacency matrix calculation
is 1000 times slower than normal calculation done by'expm',
there is no problem about the speed as long as if you don't run huge iteration or
recursion.
}
\author{
Mahiro Y (new R user)
}
